Add f9 to open in github, and <leader>g to turn it off
Yeah those are sucky options for anyone but me, but one of them at least
might be configurable soon.

Remove a lot of key filtering that was snarfing useful regular
keystrokes, so at least I can type something with a capital 'g' in it
now.

Do the git handling "in house" rather than a lot of ugly shell. Guesses
lots of things, I am okay with that.