Add a calling it just once function and key, tidy/add funcs
People? (who?) might want a "just call this once, but somehow I don't
have fugitive installed!" maybe? so make that.

Fix it running when you reenable it, so you don't have to move to get it
to show.

Put more things in to functions to make everything a big smaller for all
the testing I will never write.

Add 'abort' statements to a lot of functions, pretending I know what
they are and the repercussions of that.